requirements.txt
beautifulsoup4==4.13.3
Flask==2.2.3
Flask_Cors==3.0.10
gradio==5.20.1
langchain_chroma==0.2.2
langchain_community==0.3.19
langchain_openai==0.3.8
langchain_text_splitters==0.3.6
openai==1.66.2
python-dotenv==1.0.1
requests==2.28.1
